Several changes:
 - Dbo benchmark, resolving warning: make i signed before putting - before it
 - Added missing include to FixedSqlConnectionPool.C
 - Issue #5773: make defensive copies of parameters
   Patch provided by Bruce Toll
 - WEnvironment: Added notes about lifetime of parameter map
 - OidcUserInfoEndpoint: skip unknown scopes
 - Don't name size parameter in Dbo tutorial to avoid warnings
 - Added Wt 3.3.8 release notes
 - Don't install release notes separately, they're already part of
   the reference documentation.
 - Issue #5780: komma -> comma in doc
 - Added public vs private linkage description to some target_link_libraries calls
 - Fixes to make TechEmpower benchmarks example build on Windows
 - Json: Added missing WT_API to ParseError

+++ test/dbo/Benchmark.C
@@ -108,7 +108,7 @@ BOOST_AUTO_TEST_CASE( performance_test )
 
     p->id = i;
     p->text = text;
-    p->creation_date = Wt::WDateTime::currentDateTime().addSecs(-i * 60 * 60);
+    p->creation_date = Wt::WDateTime::currentDateTime().addSecs(-(int)i * 60 * 60);
     p->last_change_date = Wt::WDateTime::currentDateTime();
  
     for (unsigned k = 0; k < 10; ++k)
